requirements.txt
aiolimiter==1.0.0
anyio==3.7.0
APScheduler==3.6.3
async-timeout==4.0.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
exceptiongroup==1.1.1
Flask==2.3.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
python-decouple==3.8
python-telegram-bot==13.3
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
spotipy==2.23.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
Werkzeug==2.3.6
